Implement item logging
By calling grawler with the `-ilogfile <file>` flag, every item that is not an
informational text or error message is written to `<file>`. By calling
`-ilogfile -` the items are logged to stdout.

This is arguably some kind of feature creep, but this feature was requested by
users (well, one user) and I see some interesting applications, so I decided to
implement it.